requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
keyboard>=0.13.5
numpy>=1.24.0
